De-duplicating Model CreateContinuosPaymentPayload
- Added BasePaymentPayload class for overlapping payment DTO classes
- Refactored Create Payment Payload to extend BasePaymentPayload
- Refactored ContinuousPayment Payload to use BasePaymentPayload as well
- Misc. housekeeping
